remove cacher service
