+++ profiles/basic.json
@@ -33,14 +33,6 @@
       "instances": 1
     },
     {
-      "script": "svc/cacher.js",
-      "watch": true,
-      "ignore_watch": [".git", "node_modules", "public", "views"],
-      "group": "backend",
-      "exec_mode": "cluster",
-      "instances": 1
-    },
-    {
       "script": "svc/web.js",
       "watch": true,
       "ignore_watch": [".git", "node_modules", "public", "views"],
